refactor(whatsrust): extract event conversion
